fix: Account for bytes limit properly when batching records for writing (#1719)
Fixes https://github.com/cloudquery/cloudquery-issues/issues/1567
Fixes https://github.com/cloudquery/cloudquery-issues/issues/1655

The new `batch` package provides a reusable logic for batching, accounting both for the bytes & rows limits.

+++ writers/batchwriter/batchwriter_test.go
@@ -127,7 +127,7 @@ func TestBatchSize(t *testing.T) {
 		t.Run(strconv.Itoa(batchSize), func(t *testing.T) {
 			t.Parallel()
 			testClient := &testBatchClient{}
-			wr, err := New(testClient, WithBatchSize(batchSize*2))
+			wr, err := New(testClient, WithBatchSize(int64(batchSize*2)))
 			if err != nil {
 				t.Fatal(err)
 			}
